Fix remote tmux mirror file budget
